Split `Execution` types into multiple files
This is getting too big for one file.
